Move sample classes to mock namespace
