+++ tests/Mocks/SampleSubClass.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace WP_Mock\Tests\Unit\WP_Mock\Matcher;
+namespace WP_Mock\Tests\Mocks;
 
 class SampleSubClass extends SampleClass
 {
